+++ api/util.py
@@ -3,7 +3,9 @@ import re
 from api.models import OutputFormat
 
 ALL_PARTS = ['vocals', 'other', 'bass', 'drums']
-ALL_PARTS_5 = ['vocals', 'other', 'piano', 'bass', 'drums']
+ALL_PARTS_5_PIANO = ['vocals', 'other', 'piano', 'bass', 'drums']
+ALL_PARTS_5_GUITAR = ['vocals', 'other', 'guitar', 'bass', 'drums']
+ALL_PARTS_6 = ['vocals', 'other', 'piano', 'guitar', 'bass', 'drums']
 
 def get_valid_filename(s):
     """