+++ frontend/src/Constants.tsx
@@ -1,7 +1,8 @@
-export const DEFAULT_MODEL = 'spleeter';
-export const DEFAULT_MODEL_FAMILY = 'spleeter';
+export const DEFAULT_MODEL = 'bs_roformer';
+export const DEFAULT_MODEL_FAMILY = 'bs_roformer';
 export const DEFAULT_SPLEETER_MODEL = 'spleeter';
 export const DEFAULT_DEMUCS_MODEL = 'htdemucs';
+export const DEFAULT_BS_ROFORMER_MODEL = 'bs_roformer';
 
 export const LOSSY_OUTPUT_FORMATS: [number, string][] = [
   [192, '192 kbps'],