+++ frontend/src/components/Home/MusicPlayer.tsx
@@ -9,7 +9,15 @@ import { FADE_DURATION_MS } from '../../Constants';
 import { separatorLabelMap } from '../../models/Separator';
 import { SongData } from '../../models/SongData';
 import { StaticMix } from '../../models/StaticMix';
-import { AccompShortBadge, BassBadge, DrumsBadge, OriginalBadge, PianoBadge, VocalsBadge } from '../Badges';
+import {
+  AccompShortBadge,
+  BassBadge,
+  DrumsBadge,
+  GuitarBadge,
+  OriginalBadge,
+  PianoBadge,
+  VocalsBadge,
+} from '../Badges';
 import './MusicPlayer.css';
 
 interface Props {
@@ -52,11 +60,12 @@ class MusicPlayer extends React.Component<Props> {
       const accompBadge = staticMix.other ? <AccompShortBadge title="Accompaniment" /> : null;
       const bassBadge = staticMix.bass ? <BassBadge title="Bass" /> : null;
       const drumsBadge = staticMix.drums ? <DrumsBadge title="Drums" /> : null;
+      const guitarBadge = staticMix.guitar ? <GuitarBadge title="Guitar" /> : null;
       const pianoBadge = staticMix.piano ? <PianoBadge title="Piano" /> : null;
 
       audioTitleExtra = (
         <div className="badge-flex ml-2 mr-2">
-          {separatorBadge} {vocalBadge} {accompBadge} {bassBadge} {drumsBadge} {pianoBadge}
+          {separatorBadge} {vocalBadge} {accompBadge} {bassBadge} {drumsBadge} {guitarBadge} {pianoBadge}
         </div>
       );
       audioList = [